[CI/Build]: Fix Linux CUDA wheel builds (#775)
* Fix linux builds with cibuildwheel

Build is broken because torch 2.7 is not supported by CUDA 12.4 and torch
2.7 does not have distributions for manylinux_2_17_x86_64.

This commit updates the manylinux cuda image used by cibuilder to
manylinux_2_28_x86_64 which is supported by torch 2.7

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Remove cibuild configuration duplication

Configuration for cibuilder was set in pyproject.toml and also as
env variables in GitHub publish workflow.

This commit removes the configuration in workflow and centralizes
it in pyproject.toml.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Clean more disk space from runner

Running out of space when building cp312-manylinux_x86_64 wheel.
This step is to remove more space.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Refactor runner disk cleanup

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>