[Misc] GdsBackend: add internal fallback to POSIX APIs (#811)
* [Misc] GdsBackend: add internal fallback to POSIX APIs

In this change a `use_cufile` extra config variable is added to control
whether cuFile APIs will be used by GdsBackend. In addition, based on
fstype we can control the default vault of this variable and can turn
cuFile usage off.

The libcufile APIs are not always successful in implementing the POSIX
compatibility fallback. when either the cuFile own POSIX implementation
is triggered, or when GdsBackend's own POSIX implementation is used,
there expected to be at least one copy of data from main memory to GPU
buffers.

Fixes: #795

Signed-off-by: Dan Aloni <dan@kernelim.com>

* [Misc] GdsBackend: use_cufile - allow case insenstive "true" as input

Signed-off-by: Dan Aloni <dan@kernelim.com>

---------

Signed-off-by: Dan Aloni <dan@kernelim.com>